// football-coalition/src/app/models/association.model.ts
export interface Association {
  id: string;
  name: string;
  amharicName: string;
  logoPath: string;
  stampPath: string;
  stampPath2: string;
  signiturePath: string;
  phoneNumbers: string[];
  description: string;
  about: string;
  websiteLink: string;
  facebook: string;
  telegram: string;
  tikTok: string;
  primaryColor: string;
  secondaryColor: string;
  createdDate: string;
  arifPayKey: string;
}

export interface AssociationDropDown {
  id: string;
  name: string;
} 